requirements.txt
boto3
questionary
pyyaml
setuptools
appinfo